+++ libpolys/reporter/reporter.cc
@@ -353,7 +353,7 @@ void Print(const char *fmt, ...)
     char *s=(char *)omAlloc(ls+512);
 #ifdef HAVE_VSNPRINTF
     l = vsnprintf(s, ls+511, fmt, ap);
-    if ((l==-1)||(s[l]!='\0')||(l!=strlen(s)))
+    if ((l==-1)||(s[l]!='\0')||(l!=(int)strlen(s)))
     {
       printf("Print problem: l=%d, fmt=>>%s<<\n",l,fmt);
     }
